docs: fix canonical urls and sitemap
The html_baseurl is used for calculating canonical urls, and it must
point to the root of the doc tree at the docs website.  If you look at
old docs (0.16.0 for instance) you'll see the `rel="canonical"` link
points to docs.determined.ai/* instead of docs.determined.ai/latest/*,
that is, all of our canonical links were broken links.

Since we started using furo in 0.17.2, furo hasn't been setting
canonical urls at all.  For our previous (custom) theme, we configured
the canonical urls correctly (starting in 0.13.5), though the mechanism
for configuring canonical urls was different.

In newer versions of furo, canonical urls are embedded in each page, but
the html_baseurl is how they are configured, so we need that set to
include "/latest".