shared linked list
The publisher maintains a shared linked list of all events, and hands
out heads with wakeups to every stream.  Every streamer runs its filters
and writes events it cares about to the end user.

This system maxes out around 10k streamers and 400 events per second,
which feels like not very much.  It appears that the context switching
cost is dominating performance.  The easiest improvement is to run the
filters on the publishing thread, so if 10% of events are relevant, you
reduce wakeups by 90%.

Options are:

- Keep shared linked list, publisher runs filters, wakes up streamer
  with new head if necessary, then stream runs filters again.  Result:
  2x compute cost of filtering.

- Switch to per-streamer slices like []*Event[T].  Costs slightly more
  memory, but I think the system will fall over due to one websocket per
  goroutine before it falls over due to per-streamer event slices.
  Result: 8*N*R bytes of memory per event, where N is number of
  streamers and R is relevance.  For 10k streamers, 1k events per
  second, and 10% relevance, it would take 120s to build up 1G of
  memory.  That is acceptable.  Also, this solution is a lot simpler to
  implement.