passing out batches of events
- Filter within the broadcasting thread
- One wakeup per batch
- Avoid channels; use a goroutine to watch the context rather than
  converting Cond wakeups to channel wakeups with an extra context
  switch per emitted event.

This strategy is able to do about 2.2k events per second with 10k
streamers at 10% relevance.  It's a little better than the last
benchmark, but it's doing it with 250% CPU instead of 600% CPU.

The previous benchmark probably needs to be updated to not use channels
as well, since we know context switches are killing us.