reverse rbac logic
Max's idea of just gathering an integer array of every applicable userid
for each event.

This avoids the implied rbac cache invalidation logic of the other
implementations, and it multiplies the number of sub.Filter calls by the
relevance (10% in this case), which means this edition can emit nearly
20k events per second (batchSize = 100) at only 200% cpu!

UPDATE: I was double-filtering events... it's closer to 10k, but still
very good.

UPDATE: I wasn't setting users properly.... it's still 10k, but closer to
250% cpu.